// samples/nextjs/src/components/GraphQL-ConnectedDemo/query.graphql.d.ts
export declare type Maybe<T> = T | null;
export declare type Exact<T extends {
    [key: string]: unknown;
}> = {
    [K in keyof T]: T[K];
};
/** All built-in and custom scalars, mapped to their actual values */
export declare type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    JSON: any;
    /** The `Date` scalar type represents a year, month and day in accordance with the [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) standard. */
    Date: any;
    /** The `DateTime` scalar type represents a date and time. `DateTime` expects timestamps to be formatted in accordance with the [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) standard. */
    DateTime: any;
    /** The `DateTimeOffset` scalar type represents a date, time and offset from UTC. `DateTimeOffset` expects timestamps to be formatted in accordance with the [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) standard. */
    DateTimeOffset: any;
    /** The `Seconds` scalar type represents a period of time represented as the total number of seconds. */
    Seconds: any;
    /** The `Milliseconds` scalar type represents a period of time represented as the total number of milliseconds. */
    Milliseconds: any;
    Decimal: any;
};
/**
 * /sitecore/templates/Project/JssNextWeb/App Route template (ID:
 * {787584C0-A057-5876-9836-F8B3708F0CAF}). Also implements Route.
 */
export declare type AppRoute = {
    /** Page Title (ID: {F8171ECB-37C5-54EE-BFED-F689C878B6E2}). */
    pageTitle?: Maybe<TextField>;
};
// ...
export declare type GraphQlConnectedDemo = Item & {
    __typename?: 'GraphQLConnectedDemo';
    /** Security access rights for current user */
    access?: Maybe<ItemAccess>;
    children: Array<Maybe<Item>>;
    database: Scalars['String'];
    displayName: Scalars['String'];
    field?: Maybe<ItemField>;
    fields: Array<Maybe<ItemField>>;
    hasChildren: Scalars['Boolean'];
    icon: Scalars['String'];
    id: Scalars['ID'];
    insertOptions: Array<Maybe<InsertOptions>>;
    itemUri: Scalars['String'];
    language: ItemLanguage;
    name: Scalars['String'];
    parent?: Maybe<Item>;
    path: Scalars['String'];
    /** sample1 (ID: {77EC057E-91BB-576C-9488-647056849076}). */
    sample1?: Maybe<TextField>;
    /** sample2 (ID: {6654F72F-CE4D-5651-8DE3-862857A48447}). */
    sample2?: Maybe<LinkField>;
    template: ItemTemplate;
    url: Scalars['String'];
    version: Scalars['Int'];
    versions: Array<Maybe<Item>>;
    workflow: ItemWorkflow;
};
// ...
export declare type InsertOptions = {
    /** The GUID of this insert option. */
    id: Scalars['ID'];
    /** The name of this insert option. */
    name: Scalars['String'];
};
// ...
export declare type Item = {
    /** Security access rights for current user */
    access?: Maybe<ItemAccess>;
    /** Child items in the content hierarchy */
    children: Array<Maybe<Item>>;
    database: Scalars['String'];
    displayName: Scalars['String'];
    /** Single field by name or ID */
    field?: Maybe<ItemField>;
    /** All item fields. Fields can be treated as their type to get detailed info. */
    fields: Array<Maybe<ItemField>>;
    hasChildren: Scalars['Boolean'];
    icon: Scalars['String'];
    id: Scalars['ID'];
    /** Collection of item insert options. */
    insertOptions: Array<Maybe<InsertOptions>>;
    /** Uniquely identifies id, lang, version */
    itemUri: Scalars['String'];
    language: ItemLanguage;
    name: Scalars['String'];
    /** Parent in the content hierarchy. */
    parent?: Maybe<Item>;
    path: Scalars['String'];
    /** Defines item fields */
    template: ItemTemplate;
    /** Gets a URL link to the item */
    url: Scalars['String'];
    version: Scalars['Int'];
    versions: Array<Maybe<Item>>;
    workflow: ItemWorkflow;
};
// ...
export declare type ItemAccess = {
    __typename?: 'ItemAccess';
    /** Determines whether the specified branch template can be added as a child of the item. */
    canAddBranch?: Maybe<Scalars['Boolean']>;
    /** Determines whether the specified template can be added as a child of the item. */
    canAddTemplate?: Maybe<Scalars['Boolean']>;
    /** Determines whether the item can be administered (access rights set). */
    canAdmin?: Maybe<Scalars['Boolean']>;
    /** Determines whether this item can be copied to the specified destination. */
    canCopyTo?: Maybe<Scalars['Boolean']>;
    /** Determines whether a child item under this item can be created. */
    canCreate?: Maybe<Scalars['Boolean']>;
    /** Determines whether the item can be deleted. */
    canDelete?: Maybe<Scalars['Boolean']>;
    /** Determines whether the item can be duplicated. */
    canDuplicate?: Maybe<Scalars['Boolean']>;
    /** Determines whether this item can be moved to the specified destination. */
    canMoveTo?: Maybe<Scalars['Boolean']>;
    /** Determines whether the item can be read. */
    canRead?: Maybe<Scalars['Boolean']>;
    /** Determines whether the item can be read in the current language. */
    canReadLanguage?: Maybe<Scalars['Boolean']>;
    /** Determines whether this item version can be removed. */
    canRemoveVersion?: Maybe<Scalars['Boolean']>;
    /** Determines whether the item can be renamed. */
    canRename?: Maybe<Scalars['Boolean']>;
    /** Determines whether the item can be written. */
    canWrite?: Maybe<Scalars['Boolean']>;
    /** Determines whether the item can be written in the current language. */
    canWriteLanguage?: Maybe<Scalars['Boolean']>;
};
// ...
export declare type ItemField = {
    /** Whether the current user can write to this field */
    canWrite: Scalars['Boolean'];
    /** Whether the field value is a value from a language fallback item */
    containsFallbackValue: Scalars['Boolean'];
    /** Whether the value is inherited from an item clone */
    containsInheritedValue: Scalars['Boolean'];
    /** Whether the field value is a standard value */
    containsStandardValue: Scalars['Boolean'];
    /** The definition of this field on its template (field metadata) */
    definition?: Maybe<ItemTemplateField>;
    /**
     * The editable value of this field. If not in edit mode, this value is identical
     * to 'rendered'. This value may contain raw HTML; do not encode when displaying
     */
    editable?: Maybe<Scalars['String']>;
    /** The GUID of this field. */
    id: Scalars['ID'];
    /** Gets the field value rendered for use in Sitecore JSS field helper components */
    jss?: Maybe<Scalars['JSON']>;
    /** The name of this field. */
    name: Scalars['String'];
    /** The rendered value of this field (like rendered, but inline editing is disabled) */
    rendered?: Maybe<Scalars['String']>;
    /** . For rendered output to a page, use 'editable' or 'rendered' instead of this */
    value?: Maybe<Scalars['String']>;
};
export declare type ItemFieldEditableArgs = {
    fieldRendererParameters?: Maybe<Scalars['String']>;
};
export declare type ItemFieldIdArgs = {
    format?: Maybe<Scalars['String']>;
};
export declare type ItemFieldRenderedArgs = {
    fieldRendererParameters?: Maybe<Scalars['String']>;
};
export declare type ItemLanguage = {
    __typename?: 'ItemLanguage';
    displayName: Scalars['String'];
    englishName: Scalars['String'];
    icon?: Maybe<Scalars['String']>;
    name: Scalars['String'];
    nativeName: Scalars['String'];
};
export declare type ItemTemplate = InsertOptions & {
    __typename?: 'ItemTemplate';
    baseTemplates?: Maybe<Array<Maybe<ItemTemplate>>>;
    fields?: Maybe<Array<Maybe<ItemTemplateField>>>;
    id: Scalars['ID'];
    name: Scalars['String'];
    ownFields?: Maybe<Array<Maybe<ItemTemplateField>>>;
};
export declare type ItemTemplateIdArgs = {
    format?: Maybe<Scalars['String']>;
};
export declare type ItemTemplateField = {
    __typename?: 'ItemTemplateField';
    id?: Maybe<Scalars['ID']>;
    name: Scalars['String'];
    section: Scalars['String'];
    sectionSortOrder: Scalars['Int'];
    shared: Scalars['Boolean'];
    sortOrder: Scalars['Int'];
    source: Scalars['String'];
    title: Scalars['String'];
    type: Scalars['String'];
    unversioned: Scalars['Boolean'];
};
// ...
export declare type ItemWorkflow = {
    __typename?: 'ItemWorkflow';
    canEdit: Scalars['Boolean'];
    canSave: Scalars['Boolean'];
    workflowId?: Maybe<Scalars['String']>;
    workflowName?: Maybe<Scalars['String']>;
    workflowState?: Maybe<ItemWorkflowState>;
};
export declare type ItemWorkflowState = {
    __typename?: 'ItemWorkflowState';
    /** Only a final workflow state is publishable. */
    final?: Maybe<Scalars['String']>;
    /** The item ID of the workflow state item */
    id?: Maybe<Scalars['ID']>;
    name?: Maybe<Scalars['String']>;
};
// ...
export declare type LinkField = ItemField & {
    __typename?: 'LinkField';
    /** The anchor name this link points to (e.g. #foo) */
    anchor?: Maybe<Scalars['String']>;
    /** Whether the current user can write to this field */
    canWrite: Scalars['Boolean'];
    /** The CSS class on this link */
    className?: Maybe<Scalars['String']>;
    /** Whether the field value is a value from a language fallback item */
    containsFallbackValue: Scalars['Boolean'];
    /** Whether the value is inherited from an item clone */
    containsInheritedValue: Scalars['Boolean'];
    /** Whether the field value is a standard value */
    containsStandardValue: Scalars['Boolean'];
    /** The definition of this field on its template (field metadata) */
    definition?: Maybe<ItemTemplateField>;
    /** Display name of the field */
    displayName?: Maybe<Scalars['String']>;
    /**
     * The editable value of this field. If not in edit mode, this value is identical
     * to 'rendered'. This value may contain raw HTML; do not encode when displaying
     */
    editable?: Maybe<Scalars['String']>;
    /** The GUID of this field. */
    id: Scalars['ID'];
    /** Gets the field value rendered for use in Sitecore JSS field helper components */
    jss?: Maybe<Scalars['JSON']>;
    /** The type of link this is (e.g. 'external') */
    linkType?: Maybe<Scalars['String']>;
    /** The name of this field. */
    name: Scalars['String'];
    /** The query string on this link */
    queryString?: Maybe<Scalars['String']>;
    /** The rendered value of this field (like rendered, but inline editing is disabled) */
    rendered?: Maybe<Scalars['String']>;
    /** The HTML target attribute of the link (e.g. __blank) */
    target?: Maybe<Scalars['String']>;
    /** The internal item that this link targets (null for external or other link types) */
    targetItem?: Maybe<Item>;
    /** The body text of the link */
    text?: Maybe<Scalars['String']>;
    /** The URL/href of this link */
    url?: Maybe<Scalars['String']>;
    /** The raw value of this field. For rendered output to a page, use 'editable' or 'rendered' instead of this */
    value?: Maybe<Scalars['String']>;
};
// ...
export declare type TextField = ItemField & {
    __typename?: 'TextField';
    /** Whether the current user can write to this field */
    canWrite: Scalars['Boolean'];
    /** Whether the field value is a value from a language fallback item */
    containsFallbackValue: Scalars['Boolean'];
    /** Whether the value is inherited from an item clone */
    containsInheritedValue: Scalars['Boolean'];
    /** Whether the field value is a standard value */
    containsStandardValue: Scalars['Boolean'];
    /** The definition of this field on its template (field metadata) */
    definition?: Maybe<ItemTemplateField>;
    /** Display name of the field */
    displayName?: Maybe<Scalars['String']>;
    /**
     * The editable value of this field. If not in edit mode, this value is identical
     * to 'rendered'. This value may contain raw HTML; do not encode when displaying
     */
    editable?: Maybe<Scalars['String']>;
    /** The GUID of this field. */
    id: Scalars['ID'];
    /** Gets the field value rendered for use in Sitecore JSS field helper components */
    jss?: Maybe<Scalars['JSON']>;
    /** The name of this field. */
    name: Scalars['String'];
    /** The rendered value of this field (like rendered, but inline editing is disabled) */
    rendered?: Maybe<Scalars['String']>;
    /** The raw value of this field. For rendered output to a page, use 'editable' or 'rendered' instead of this */
    value?: Maybe<Scalars['String']>;
};
// ...
export declare const ConnectedDemoQueryDocument: import("graphql").DocumentNode;